requirements.txt
channels==4.0.0
daphne==4.1.0
Django==5.0.3
djangorestframework==3.15.1
pillow==10.2.0


